requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
SQLAlchemy==2.0.36
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
bcrypt==3.2.2
requests==2.32.3
python-dotenv==1.0.1   # opcional, se quiser carregar .env
email-validator
